Added multi const support to the interpreter
